+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 1.1.2
 
 * Demo Screenshot added
+
+## 1.1.3
+
+* LoadingButtonTypes renamed from Raised to Elevated and Flat to Text
+* loadingWidget is now nullable by default it will show default loading widget